refactor(flydsl): split grouped bf16 GEMM out of the dense bf16 kernel
gemm/gemm_bf16_kernel.py mixed the dense bf16 GEMM with the variable-K
grouped wgrad kernel. Move the grouped half to its own module so the
bf16 pair mirrors the fp8/mxfp8/mxfp4 layout under grouped_gemm/.

- new grouped_gemm/grouped_gemm_bf16_kernel.py holds the variable-K
  tile, its compile cache and the launcher
- rename to match the grouped_gemm/ naming scheme:
  gemm_bf16_variable_k_tile      -> grouped_gemm_bf16_variable_k_tile
  grouped_gemm_variable_k_bf16   -> grouped_gemm_bf16_variable_k_flydsl_kernel
- move _i64 to utils/gemm_helper.py: it is used only by the grouped
  side now, and gemm_helper.py is where cross-directory private
  helpers already live (_readfirstlane_i32, _robust_time)
- _make_shared_storage stays in gemm/ and is imported back by the
  grouped module, same as grouped_gemm_mxfp4_kernel.py already does

Pure relocation, no kernel logic changed. Verified bit-identical
results on gfx950 for grouped / trans_c / masked_k / dense NN / dense TN.
